refactor(Home): refacto page Home add Logo
