+++ src/index.css
@@ -14,3 +14,8 @@
 .colorr:hover span {
   color: rgb(202, 138, 4);
 }
+
+.addMore-img {
+  filter: brightness(0) saturate(100%) invert(51%) sepia(100%) saturate(409%)
+    hue-rotate(3deg) brightness(89%) contrast(101%);
+}
