+++ src/shared/components/Header.tsx
@@ -1,8 +1,5 @@
-type Language = {
-  id: number;
-  color: string;
-  name: string;
-};
+import logo from "../../assets/logo.svg";
+import Language from "../../types/Languages";
 
 const languages: Language[] = [
   { id: 1, color: "bg-yellow-300", name: "javascript" },
@@ -14,17 +11,14 @@ const languages: Language[] = [
 
 const Header = () => {
   return (
-    <header className="w-1/2 flex cursor-default my-10">
-      <div className="flex flex-col gap-3">
-        <h1 className="text-6xl font-bold text-transparent bg-clip-text bg-gradient-to-tl from-yellow-900 via-yellow-600 to-yellow-400">
+    <header className="w-full flex flex-wrap-reverse justify-between items-center">
+      <div className="w-1/2">
+        <h1 className="text-9xl font-bold text-transparent bg-clip-text bg-gradient-to-tl from-yellow-900 via-yellow-600 to-yellow-400">
           dePo.
         </h1>
-        <p className="text-3xl font-semibold text-slate-50">
-          El repositorio en el que puedes encontrar los comandos, metodos de
-          arrays y de strings, mas usados al desarrollar con las tecnologias y
-          lenguajes mas famosos del mercado.
+        <p className="text-4xl font-extrabold text-slate-50 my-4">
+          Metodos y comandos a un click de tu mouse.
         </p>
-        <p className="text-2xl">Preparate para navegar entre lenguajes.</p>
         <p>
           {languages.map((language: Language) => (
             <span
@@ -35,6 +29,12 @@ const Header = () => {
             </span>
           ))}
         </p>
+      </div>
+      <div className="w-1/2 relative">
+        <div className="w-[18rem] h-[18rem] absolute left-10 rounded-full bg-gradient-to-r from-yellow-200 to-yellow-500 blur-3xl opacity-40"></div>
+        <div className="w-[18rem] h-[18rem] rounded-full absolute right-10 bg-gradient-to-r from-red-800 via-yellow-600 to-yellow-500 blur-3xl opacity-40"></div>
+        <div className="w-[18rem] h-[18rem] bg-amber-200 rounded-full absolute bottom-[-3rem] left-1/4 bg-gradient-to-tl from-yellow-900 via-yellow-600 to-yellow-400 blur-3xl opacity-40"></div>
+        <img className="addMore-img mx-auto" src={logo} alt="logo" />
       </div>
     </header>
   );
