+++ src/shared/components/Navbar.tsx
@@ -2,8 +2,8 @@ import { NavLink } from "react-router-dom";
 
 const Navbar = () => {
   return (
-    <nav className="h-[4rem] bg-[#1e1e20]">
-      <div className="h-[5rem] max-w-6xl flex justify-between items-center fixed bg-[#1e1e20] top-0 right-0 left-0 mx-auto mb-10 px-7 z-20">
+    <nav className="h-[5rem] w-screen bg-[#1e1e20] sticky top-0 z-20 mb-32">
+      <div className="flex h-[5rem] max-w-7xl justify-between items-center bg-[#1e1e20] p-6 mx-auto">
         <span className="text-3xl font-bold text-transparent bg-clip-text bg-gradient-to-tl from-yellow-900 via-yellow-600 to-yellow-400 cursor-pointer">
           <NavLink to={"/"}>dePo</NavLink>
         </span>
@@ -13,7 +13,7 @@ const Navbar = () => {
             className={({ isActive }) =>
               !isActive
                 ? "cursor-pointer font-medium text-[#b1b1b1] hover:text-yellow-600"
-                : "text-yellow-600"
+                : "text-yellow-600 font-medium"
             }
           >
             Inicio
@@ -23,7 +23,7 @@ const Navbar = () => {
             className={({ isActive }) =>
               !isActive
                 ? "cursor-pointer font-medium text-[#b1b1b1] hover:text-yellow-600"
-                : "text-yellow-600"
+                : "text-yellow-600 font-medium"
             }
           >
             arrays
@@ -33,7 +33,7 @@ const Navbar = () => {
             className={({ isActive }) =>
               !isActive
                 ? "cursor-pointer font-medium text-[#b1b1b1] hover:text-yellow-600"
-                : "text-yellow-600"
+                : "text-yellow-600 font-medium"
             }
           >
             strings
@@ -43,7 +43,7 @@ const Navbar = () => {
             className={({ isActive }) =>
               !isActive
                 ? "cursor-pointer font-medium text-[#b1b1b1] hover:text-yellow-600"
-                : "text-yellow-600"
+                : "text-yellow-600 font-medium"
             }
           >
             comandos
