+++ src/types/Languages.ts
@@ -1,0 +1,7 @@
+type Language = {
+  id: number;
+  color: string;
+  name: string;
+};
+
+export default Language;
